ThemeChooser: new launcher icon and removal of annoying warning
The new launcher icon was taken from CyanogenMod 12.0, with the "Cid"
character removed. All part of the "modernization" of Gingerbread.

The warning that was removed was the "this theme is missing assets for
your device's screen size" message. This usually happened with themes
that were created under Windows, and since themes that generate this
warning usually work just fine, this warning was mostly annoying (and
was usually ignored by most people). Therefore, the warning is no longer
displayed and the theme is always applied.

Note that the "this theme was improperly compiled" warning is still in
place, since this indicates an error in the theme that needs to be
corrected.